Completed Challenge 12

+++ Utilities.h
@@ -447,7 +447,7 @@ uint8_t OracleP11(uint8_t* buff, uint8_t len) {
 #endif
 
 #ifdef CH_2_12
-uint16_t OracleP12(uint8_t* buff, uint8_t len) {
+uint16_t OracleP12(uint8_t* buff, uint16_t len) {
   uint8_t Iv[16];
   fillRandom(Iv, 16);
   uint8_t myBuff[512];
